refactor: Playwright config
