+++ swapstreet/playwright.config.ts
@@ -1,14 +1,6 @@
 import { defineConfig, devices } from '@playwright/test';
 
 /**
- * Read environment variables from file.
- * https://github.com/motdotla/dotenv
- */
-// import dotenv from 'dotenv';
-// import path from 'path';
-// dotenv.config({ path: path.resolve(__dirname, '.env') });
-
-/**
  * See https://playwright.dev/docs/test-configuration.
  */
 export default defineConfig({
@@ -34,9 +26,10 @@ export default defineConfig({
     actionTimeout: 10_000,
   },
 
-    /* Run local dev server before starting the tests (helpful in CI) */
+  /* Run local dev server before starting the tests (helpful in CI) */
   webServer: {
-    command: 'npm --prefix swapstreet ci && npm --prefix swapstreet run dev',
+    // when running inside swapstreet, start dev server directly
+    command: 'npm ci && npm run dev',
     url: 'http://localhost:3000',
     reuseExistingServer: true,
     timeout: 120_000,
@@ -44,16 +37,16 @@ export default defineConfig({
 
   /* Configure projects for browser + device combinations (desktop / tablet / mobile) */
   projects: [
-      // Desktop browsers
-      { name: "chromium-desktop", use: { ...devices["Desktop Chrome"] } },
-      { name: "firefox-desktop", use: { ...devices["Desktop Firefox"] } },
-      { name: "webkit-desktop", use: { ...devices["Desktop Safari"] } },
-  
-      // Tablet
-      { name: "ipad", use: { ...devices["iPad (gen 7)"] } },
-  
-      // Mobile
-      { name: "iphone", use: { ...devices["iPhone 13"] } },
-      { name: "pixel-5", use: { ...devices["Pixel 5"] } },
-    ],
+    // Desktop browsers
+    { name: 'chromium-desktop', use: { ...devices['Desktop Chrome'] } },
+    { name: 'firefox-desktop', use: { ...devices['Desktop Firefox'] } },
+    { name: 'webkit-desktop', use: { ...devices['Desktop Safari'] } },
+
+    // Tablet
+    { name: 'ipad', use: { ...devices['iPad (gen 7)'] } },
+
+    // Mobile
+    { name: 'iphone', use: { ...devices['iPhone 13'] } },
+    { name: 'pixel-5', use: { ...devices['Pixel 5'] } },
+  ],
 });
